OCR/Tesseract/requirements.txt
django==1.11
djangorestframework

# packages for tesseract wrapper: pyocr
Pillow
pyocr
